Style the emergency banner.
We added a sass/layout/_layouts.scss file for layout rules, and a sass/blocks/_blocks.scss for 'block level' rules (ie. a block being a self contained element that can be placed). For now we add our own variables at the top of sass/style.scss. The styling of the emergency banner uses a bit of magic with pseudo elements for spreading the color across the page and for the arrow.