Create home page widget.
Create a custom WP widget that can be used to display the blocks on the home page. This attempts to find a balance between allowing them some control, but not forcing them to write HTML. It attempts to cater for the existing cases in the design, apart from the main large block which will be a separate task as a lot of time was spent on this already.